ReadOnly nodes should not require preset saves
This change introduces ReadOnly "category picker" nodes into crafting. A category picker is a node which creates filtered lists of child nodes (ie "mass wall" vs "masonry wall"). Unlike other nodes, ReadOnly nodes are not presettable. They do not get dirtied or created by the user. So if a node has a ReadOnly node in one of its pins, it also has to record the writeable node that is attached to the ReadOnly, otherwise the ReadOnly node is a dead end.

In consideration of this, node pins have been refactored to accept a string of child nodes consisting of zero or more ReadOnly nodes followed by zero or one writeable node.

When a sequence ends with a writeable node, as is the case for all user-defined presets and complete assemblies and layers, the sequence is considered fully qualified and can be restored as is. When a sequence ends in a ReadOnly node, as is the case for "default assemblies" defined in the DDL script, we follow the chain of default children (also defined in the script) until we find a writeable node to terminate with.

This allows a crafting tree to be structured as:

My Wall->Masonry Category->Clay Brick Subcategory->My Clay Brick Layer

In this case, the My Wall node (which is writeable) has one direct child, Masonry Category, which is ReadOnly and that node subsequently has Clay Brick Subcategory assigned (which is also ReadOnly). The subcategory has My Clay Brick Layer attached, which is writeable.

With the child sequence system, instead of just pointing to Masonry Category and following its default children, we make a 3 member sequence of the category, the subcategory and My Clay Brick Layer and attach all of them to My Wall, which now has a complete record of how it is laid out.

Code notes:

* Creating a node from a preset now needs to know whether it should fill out the "default" children of presets that end in ReadOnly nodes (as is the case for starting a new crafting session) or whether they should leave ReadOnly nodes childless and let the sequence assign children (as is the case for editing an existing assembly)

* When making a spec out of a tree, we only consider the last child in a sequence. ReadOnly nodes don't have any pertinent properties for making assemblies. They're just used to group things during crafting

* Bug fix: beams and columns had not been updated to use the unified Extents dimension naming

Notion:
https://www.notion.so/Support-non-presettable-pass-through-category-subcategory-nodes-in-DDL-c23013def9b341799749a2041d5b5a4a